requirements.txt
boto3 
pyyaml=5.4.1
cfn-lint 
